Title and version
Trying to solve the first problem, I change the title of the package from "Nowcasting Analysis and Create Real-Time Data Basis" to "Nowcast Analysis and Create Real-Time Data Basis" and change the package version from 0.0.9 to 0.0.90.

To the second problem I remove the redundant packages and allow installation for R 3.3.1 version.